+++ src/DependencyInjection/RekalogikaMapperExtension.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\DependencyInjection;
 
-use ReflectionNamedType;
 use Rekalogika\Mapper\Attribute\AsObjectMapper;
 use Rekalogika\Mapper\Attribute\AsPropertyMapper;
 use Rekalogika\Mapper\Exception\LogicException;
@@ -74,7 +73,6 @@ final class RekalogikaMapperExtension extends Extension
         AsPropertyMapper $attribute,
         \ReflectionMethod $reflector,
     ): void {
-        /** @var array{sourceClass:class-string,targetClass:class-string,method:string} */
         $tagAttributes = [];
 
         // get the AsPropertyMapper attribute attached to the class
@@ -116,17 +114,41 @@ final class RekalogikaMapperExtension extends Extension
         $firstParameter = $parameters[0] ?? null;
         $type = $firstParameter?->getType();
 
-        if ($type === null || !$type instanceof \ReflectionNamedType) {
-            throw new LogicException(
-                \sprintf(
-                    'Unable to determine the source class for property mapper service "%s", method "%s".',
-                    $definition->getClass() ?? '?',
-                    $reflector->getName(),
-                ),
-            );
-        }
+        if ($type === null) {
+            throw new LogicException(\sprintf(
+                'Cannot set up property mapper, the type of the first argument cannot be determined. Service ID "%s", method "%s".',
+                $definition->getClass() ?? 'unknown',
+                $reflector->getName(),
+            ));
+        } elseif ($type instanceof \ReflectionNamedType) {
+            $sourceClasses = [$type->getName()];
+        } elseif ($type instanceof \ReflectionUnionType) {
+            $sourceClasses = [];
 
-        $tagAttributes['sourceClass'] = $type->getName();
+            foreach ($type->getTypes() as $type) {
+                if ($type instanceof \ReflectionIntersectionType) {
+                    throw new LogicException(\sprintf(
+                        'Cannot set up property mapper, the type of the first argument contains an intersection type, which is not supported. Service ID "%s", method "%s".',
+                        $definition->getClass() ?? '?',
+                        $reflector->getName(),
+                    ));
+                } elseif (!$type instanceof \ReflectionNamedType) {
+                    throw new LogicException(\sprintf(
+                        'Cannot set up property mapper, the type of the first argument contains a non-named type, which is not supported. Service ID "%s", method "%s".',
+                        $definition->getClass() ?? '?',
+                        $reflector->getName(),
+                    ));
+                }
+
+                $sourceClasses[] = $type->getName();
+            }
+        } else {
+            throw new LogicException(\sprintf(
+                'Cannot set up property mapper, the type of the first argument is unsupported. Service ID "%s", method "%s".',
+                $definition->getClass() ?? '?',
+                $reflector->getName(),
+            ));
+        }
 
         // if the property is missing, assume it is the same as the
         // method name
@@ -149,19 +171,6 @@ final class RekalogikaMapperExtension extends Extension
             $name = lcfirst($name);
 
             $tagAttributes['property'] = $name;
-        }
-
-        // if the sourceClass is not a class, throw an exception
-
-        if (!class_exists($tagAttributes['sourceClass'])) {
-            throw new \LogicException(
-                \sprintf(
-                    'Source class "%s" for property mapper service "%s", method "%s" does not exist.',
-                    $tagAttributes['sourceClass'],
-                    $definition->getClass() ?? '?',
-                    $reflector->getName(),
-                ),
-            );
         }
 
         // if the targetClass is still missing, throw an exception
@@ -191,7 +200,12 @@ final class RekalogikaMapperExtension extends Extension
 
         // finally
 
-        $definition->addTag('rekalogika.mapper.property_mapper', $tagAttributes);
+        foreach ($sourceClasses as $sourceClass) {
+            $definition->addTag('rekalogika.mapper.property_mapper', [
+                ...$tagAttributes,
+                'sourceClass' => $sourceClass,
+            ]);
+        }
     }
 
     private function objectMapperConfigurator(